Clean up and separate namespaces in tests
